Add files via upload
This is the first task I've completed as an intern at the Sparks Foundation. The given raw data has been used to predict the score of a student that studies for 9.25 hours a day. It is an example of prediction using supervised ML. The prediction has been done using a simple linear regression model.